examples/a2a_langgraph_crewai/pyproject.toml
[project]
name = "a2a-langgraph-crewai"
version = "0.1.0"
description = "A2A demo: LangGraph orchestrator calls CrewAI specialist over JSON-RPC (a2a-sdk)."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "crewai[litellm]>=1.10.1",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langgraph-prebuilt>=1.0.8",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

# Script-style layout: same directory as pyproject.toml
[tool.setuptools]
py-modules = ["a2a_reply", "crew_a2a_server", "langgraph_a2a_server", "demo_client"]
